pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "lec"
version = "0.1.0"
description = "Living Evidence Composite (LEC) Pipeline - Phase 0 Bronze"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "click>=8.0",
    "pydantic>=2.0",
    "jsonschema>=4.0",
    "pyarrow>=12.0",
    "pandas>=2.0",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
]
postgres = [
    "psycopg2-binary>=2.9",
    "sqlalchemy>=2.0",
]

[project.scripts]
lec = "lec.cli:main"

[tool.setuptools.packages.find]
where = ["src"]
